Move User and UserInDB classes from models to schemas due to those classes are also used by internal functions, not only the API
